features.value_bounds -> features.util.value_bounds
Move this utility function with other bounds tests.
